requirements.txt
opencv-python==4.6.0.66
Pillow==9.2.0
face_recognition==1.3.0

SQLAlchemy~=2.0.30
numpy~=1.23.5
torch~=1.13.1
torchvision~=0.14.1
tqdm~=4.64.1
tensorboardX~=2.5.1
easydict~=1.10